requirements.txt
llama-index
llama-index-readers-docling
llama-index-embeddings-ollama
llama-index-llms-ollama
llama-index-vector-stores-postgres
llama-index-storage-docstore-postgres
llama-index-storage-docstore-postgres
python-dotenv
psycopg2-binary
crewai
crewai-tools